SOLIDER-REID-Lightning/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
lightning>=2.6.0  # ✅ Updated to latest Lightning (formerly pytorch-lightning)
torchmetrics>=1.0.0

# Configuration
hydra-core>=1.3.0
omegaconf>=2.3.0
yacs>=0.1.8

# Data processing
timm>=0.6.0
opencv-python>=4.5.0
pillow>=8.0.0
scikit-learn>=1.0.0

# Metrics & Evaluation
faiss-gpu>=1.7.0  # or faiss-cpu if no GPU
scipy>=1.7.0

# Logging & Visualization
wandb>=0.13.0

# Utilities
tqdm>=4.62.0
pyyaml>=5.4.0
numpy>=1.21.0
